+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * build(deps): bump github.com/stretchr/testify from 1.7.0 to 1.7.2
 * feat(log-drains): The `LogDrainsAddonList` now returns the list of log drains [#246](https://github.com/Scalingo/go-scalingo/pull/246)
 * feat(log-drains): Cleanup the LogDrain struct [#246](https://github.com/Scalingo/go-scalingo/pull/246)
+* refactor(domain): deprecate DomainsUpdate
 
 ## 4.15.1